Pull ECS tasks into thier own file
